pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gugs"
version = "0.1.0"
description = "Github Username Gravity Simulation - Create animated GIFs with particles forming text"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Nicolo' Massari", email = "nmassari@ethz.ch"}
]
dependencies = [
    "numpy>=1.21.0",
    "matplotlib>=3.4.0",
    "pillow>=8.2.0"
]

[project.scripts]
gugs = "gugs.__main__:main"

[tool.setuptools]
packages = ["gugs"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
gugs = ["../liberation-mono/*.ttf"]